backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database & ORM
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
email-validator==2.1.0

# Pydantic & Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# AI & LLM
google-generativeai==0.3.2
openai==1.3.7

# Vector Search & Embeddings
pgvector==0.2.4
sentence-transformers==2.2.2
numpy==1.24.3

# Cache & Queue
redis==5.0.1
celery==5.3.4
rq==1.15.1

# File Processing
pypdf2==3.0.1
python-docx==1.1.0
pillow==10.1.0

# HTTP & Async
httpx==0.25.2
aiofiles==23.2.1

# Monitoring & Observability
prometheus-client==0.19.0
sentry-sdk[fastapi]==1.38.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0

# Search & Text Processing
whoosh==2.7.4
unidecode==1.3.7

# Background Tasks
apscheduler==3.10.4

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
